pyproject.toml
[project]
name = "kaggle-home-prices"
version = "0.1.0"
description = ""
authors = [
    {name = "Vlad Mashtalier",email = "vmashtaler5@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "optuna (>=4.5.0,<5.0.0)",
    "mlflow (>=3.4.0,<4.0.0)",
    "xgboost (>=3.0.5,<4.0.0)",
    "notebook (>=7.4.7,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
